fix(support): move evidence helper out of server action
